0217-contains-duplicate incomplete solution

+++ problems/0217-contains-duplicate/contains-duplicate.h
@@ -4,7 +4,7 @@
 #include <stdbool.h>
 #include <stddef.h>                
 
-bool containsDuplicate(int *nums, int numsSize);
+bool containsDuplicate(int *nums, size_t numsSize);
 
 void solve();
